dd kernel refactor in progress

+++ include/pirate/DedispersionKernel.hpp
@@ -244,8 +244,8 @@ struct dd_kernel
     // The low-level cuda kernel is called as:
     //
     // void dd_kernel(
-    //     void *inbuf, int act_istride32, int amb_istride32, long beam_istride32,
-    //     void *outbuf, int act_ostride32, int amb_ostride32, long beam_ostride32,
+    //     void *inbuf, long beam_istride32, int amb_istride32, int act_istride32,
+    //     void *outbuf, long beam_ostride32, int amb_ostride32, int act_ostride32,
     //     void *pstate, int ntime, long rb_pos
     // );
     //
@@ -267,7 +267,7 @@ struct dd_kernel
     //   -  The kernel is launched with {32, warps_per_threadblock} warps
     //      and {namb, nbeams} blocks.
 
-    void (*cuda_kernel)(void *, int, int, long, void *, int, int, long, void *, int, long) = nullptr;
+    void (*cuda_kernel)(void *, long, int, int, void *, long, int, int, void *, int, long) = nullptr;
 
     int shmem_nbytes = 0;
     int warps_per_threadblock = 0;
